backend/pyproject.toml
[tool.poetry]
name = "ratereel"
version = "0.1.0"
description = ""
authors = ["GoshkaLP <rybkin.georgy@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
